requirements.txt
# Core dependencies for development and deployment scripts
snowflake-cli>=2.0.0
pyyaml>=6.0
pandas>=2.0.0
streamlit>=1.28.0

# Development tools
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0

# Optional: For local development testing
snowflake-snowpark-python>=1.11.0
plotly>=5.0.0 